# docs/swagger.yaml
definitions:
...
  domain.Patient:
    properties:
      created_at:
        type: string
      id:
        type: integer
      identity_number:
        type: string
      name:
        type: string
      surname:
        type: string
    required:
    - created_at
    - identity_number
    - name
    - surname
    type: object
  web.errorResponse:
    properties:
      message:
        type: string
      status:
        type: string
      status_code:
        type: integer
      time_stamp:
        type: string
    type: object
  web.response:
    properties:
      data: {}
    type: object
info:
  contact:
    name: Marcelo Ramos
    url: https://github.com/rows15
  description: This API handle appointments, dentists and patients for dental clinic
    system.
  license:
    name: Apache 2.0
    url: https://www.apache.org/license/LICENSE-2.0.html
  termsOfService: https://www.linkedin.com/in/rows15
  title: Dental Clinic API
  version: "1.0"
paths:
...
  /patients:
    get:
      consumes:
      - application/json
      description: get all patients from db.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Patient'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.errorResponse'
      security:
      - SECRET_TOKEN: []
      summary: List all patients
      tags:
      - Patients
    post:
      consumes:
      - application/json
      description: Create a new patient by request body.
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.Patient'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Patient'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.errorResponse'
      security:
      - SECRET_TOKEN: []
      summary: Create a new patient
      tags:
      - Patients
    put:
      consumes:
      - application/json
      description: Update an entire patient by ID.
      parameters:
      - description: Patient ID
        in: path
        name: id
        required: true
        type: integer
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.Patient'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Patient'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.errorResponse'
      security:
      - SECRET_TOKEN: []
      summary: Update an entire patient by ID
      tags:
      - Patients
  /patients/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a patient by ID
      parameters:
      - description: Patient ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
      security:
      - SECRET_TOKEN: []
      summary: Delete a patient by ID
      tags:
      - Patients
    get:
      consumes:
      - application/json
      description: get a patient by a provided ID.
      parameters:
      - description: Patient ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Patient'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
      security:
      - SECRET_TOKEN: []
      summary: Get a patient by an ID
      tags:
      - Patients
    patch:
      consumes:
      - application/json
      description: Update fields from a patient
      parameters:
      - description: Patient ID
        in: path
        name: id
        required: true
        type: integer
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.Patient'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Patient'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.errorResponse'
      security:
      - SECRET_TOKEN: []
      summary: Update fields from a patient
      tags:
      - Patients
securityDefinitions:
  SECRET_TOKEN:
    in: header
    name: SECRET_TOKEN
    type: apiKey
swagger: "2.0"